Game loop

+++ game/src/data/components.rs
@@ -1,9 +1,9 @@
 use std::collections::HashMap;
 
 pub struct Components<'a> {
-    pub current_entities: u32,
-    pub descriptions: HashMap<u32, &'a str>,
-    pub destinations: HashMap<u32, u32>,
-    pub locations: HashMap<u32, Vec<u32>>,
-    pub names: HashMap<u32, &'a str>,
+    pub current_entities: usize,
+    pub descriptions: HashMap<usize, &'a str>,
+    pub destinations: HashMap<usize, usize>,
+    pub locations: HashMap<usize, Vec<usize>>,
+    pub names: HashMap<usize, &'a str>,
 }
